+++ changelog.md
@@ -1,3 +1,12 @@
+## v0.9.3
+* You can now use negative values in source code
+* Implemented SourceLocation, although it is not yet visible anywhere
+* Replaced ExtendedMath.rnt with StandardMath.h functions
+	* Sine()
+	* Cosine()
+* Added functions to StandardLibrary.h
+	* Input()
+
 ## v0.9.2
 * Attempted to properly implement lazy evaluation
 * Implemented decimal value support
@@ -18,6 +27,5 @@
 	* Factorial()
 	* Power()
 	* Sine() Unfinished for now, only works right for certain values
-* Added
 ## Log keeping begins here
 If you need to know about the progress in earlier commits, you're going to have to go on commit messages, or manually dig through the code changes.